record input, save it and make playthroughs replayable

+++ draw_utils.go
@@ -55,11 +55,11 @@ func SubImage(screen *ebiten.Image, r image.Rectangle) *ebiten.Image {
 }
 
 func DrawPixel(screen *ebiten.Image, pt Pt, color color.Color) {
-	size := 2
+	size := int64(2)
 	m := screen.Bounds().Min
 	for ax := pt.X - size; ax <= pt.X+size; ax++ {
 		for ay := pt.Y - size; ay <= pt.Y+size; ay++ {
-			screen.Set(m.X+ax, m.Y+ay, color)
+			screen.Set(m.X+int(ax), m.Y+int(ay), color)
 		}
 	}
 }
