+++ geometry_test.go
@@ -14,13 +14,13 @@ func BenchmarkMoveRect(b *testing.B) {
 
 	obstacles := make([]Rectangle, 30)
 	for i := range obstacles {
-		x := i % 5
-		y := i / 5
+		x := int64(i % 5)
+		y := int64(i / 5)
 		obstacles[i].Corner1 = Pt{x * 120, y * 120}
 		obstacles[i].Corner2 = obstacles[i].Corner1.Plus(brickSize)
 	}
 
-	nMaxPixels := 100
+	nMaxPixels := int64(100)
 
 	for b.Loop() {
 		// If I put the code here directly instead of wrapped into a function,
@@ -29,7 +29,7 @@ func BenchmarkMoveRect(b *testing.B) {
 	}
 }
 
-func f(r Rectangle, targetPos Pt, nMaxPixels int, obstacles []Rectangle) Rectangle {
+func f(r Rectangle, targetPos Pt, nMaxPixels int64, obstacles []Rectangle) Rectangle {
 	r, nMaxPixels = MoveRect(r, targetPos, nMaxPixels, obstacles)
 	r, nMaxPixels = MoveRect(r, Pt{targetPos.X, r.Corner1.Y}, nMaxPixels, obstacles)
 	r, nMaxPixels = MoveRect(r, Pt{r.Corner1.X, targetPos.Y}, nMaxPixels, obstacles)
