+++ point.go
@@ -1,11 +1,11 @@
 package main
 
 type Pt struct {
-	X int
-	Y int
+	X int64
+	Y int64
 }
 
-func (p Pt) SquaredDistTo(other Pt) int {
+func (p Pt) SquaredDistTo(other Pt) int64 {
 	return p.To(other).SquaredLen()
 }
 
@@ -27,15 +27,15 @@ func (p *Pt) Subtract(other Pt) {
 	p.Y = p.Y - other.Y
 }
 
-func (p Pt) Times(multiply int) Pt {
+func (p Pt) Times(multiply int64) Pt {
 	return Pt{p.X * multiply, p.Y * multiply}
 }
 
-func (p Pt) DivBy(divide int) Pt {
+func (p Pt) DivBy(divide int64) Pt {
 	return Pt{p.X / divide, p.Y / divide}
 }
 
-func (p Pt) SquaredLen() int {
+func (p Pt) SquaredLen() int64 {
 	return p.X*p.X + p.Y*p.Y
 }
 
@@ -43,6 +43,6 @@ func (p Pt) To(other Pt) Pt {
 	return Pt{other.X - p.X, other.Y - p.Y}
 }
 
-func (p Pt) Dot(other Pt) int {
+func (p Pt) Dot(other Pt) int64 {
 	return p.X*other.X + p.Y*other.Y
 }
